Avoid setting app config variables as module attributes at compile time
Module attributes were being set at compilation time from the application
configuration. This is normally not a problem when done within an app,
because mix triggers a full app recompile if a config file changes.

For dependencies, mix does not attempt to rebuild dependencies when the
config file changes if it has already been compiled.