Asynchronously consume messages.
We can control how many unacked messages rabbit will send us by changing the `consumer_count` configuration but the code still processed those messages serially. This change spawns a new process for each message we receive.

This has a pretty considerable impact on the speeed with which we can process messages. We may need to keep an eye out for database connection depletion issues though. I think the worst that will happen is we get a timeout when trying to get a connection from the connection pool and the message will be requeued and retried.