Share a persistent connection between all queue subscribers.
Since queue subscriptions require a persistent connection, this change uses a separate connection instead of reusing one from the connection pool. This circumvents the issue we've seen where we resubscribe on an overflow connection that is closed after it is returned to the pool. By having the supervisor manage this connection we can employ the `rest_for_one` strategy so that if the connection dies all workers will be killed as well and then restarted after the connection restarts.